requirements.txt
asgiref==3.6.0
Django==4.0.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djongo==1.3.6
pymongo==3.12.3
pytz==2022.7
sqlparse==0.2.4
